+++ src/mqtt_client.rs
@@ -56,7 +56,7 @@ pub struct MqttClient<
     const MSG_SIZE: usize,
     const MSG_COUNT: usize,
 > {
-    pub(crate) network: InterfaceHolder<TcpStack>,
+    pub(crate) network: InterfaceHolder<TcpStack, MSG_SIZE>,
     clock: Clock,
     session_state: SessionState<Clock, MSG_SIZE, MSG_COUNT>,
     connection_state: StateMachine<Context>,
@@ -125,6 +125,9 @@ where
 
             _ => {}
         }
+
+        // Attempt to finish any pending packets.
+        self.network.finish_write()?;
 
         self.handle_timers()?;
 
@@ -203,6 +206,11 @@ where
             return Err(Error::NotReady);
         }
 
+        // We can't subscribe if there's a pending write in the network.
+        if self.network.has_pending_write() {
+            return Err(Error::NotReady);
+        }
+
         let packet_id = self.session_state.get_packet_identifier();
 
         let mut buffer: [u8; MSG_SIZE] = [0; MSG_SIZE];
@@ -248,6 +256,12 @@ where
     /// # Returns
     /// True if the client is able to service QoS 1 requests.
     pub fn can_publish(&self, qos: QoS) -> bool {
+        // We cannot publish if there's a pending write in the network stack. That message must be
+        // completed first.
+        if self.network.has_pending_write() {
+            return false;
+        }
+
         self.session_state.can_publish(qos)
     }
 
@@ -273,12 +287,13 @@ where
         retain: Retain,
         properties: &[Property],
     ) -> Result<(), Error<TcpStack::Error>> {
-        // TODO: QoS 2 support.
-        assert!(qos != QoS::ExactlyOnce);
-
         // If we are not yet connected to the broker, we can't transmit a message.
         if self.is_connected() == false {
             return Ok(());
+        }
+
+        if !self.can_publish(qos) {
+            return Err(Error::NotReady);
         }
 
         debug!(
@@ -353,6 +368,11 @@ where
 
         // Replay QoS 1 messages
         for key in self.session_state.pending_publish_ordering.iter() {
+            // If the network stack cannot send another message, do not attempt to send one.
+            if self.network.has_pending_write() {
+                break;
+            }
+
             let message = self.session_state.pending_publish.get(&key).unwrap();
             self.network.write(message)?;
         }
@@ -436,6 +456,11 @@ where
     fn handle_timers(&mut self) -> Result<(), Error<TcpStack::Error>> {
         // If we are not connected, there's no session state to manage.
         if self.connection_state.state() != &States::Active {
+            return Ok(());
+        }
+
+        // If there's a pending write, we can't send a ping no matter if it is due.
+        if self.network.has_pending_write() {
             return Ok(());
         }